fix code style issues in TimestampFormatter.scala
